# presenceBackend/src/main/resources/openAPI.yaml
openapi: 3.0.0
info:
  title: HEI Absence Management API
  description: API for managing student absences at HEI
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
paths:
  /students:
    get:
      summary: Get all students
      responses:
        '200':
          description: A list of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
    post:
      summary: Create a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Student'
      responses:
        '201':
          description: Student created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
  /students/{id}:
    get:
      summary: Get a student by ID
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the student to retrieve
          schema:
            type: string
      responses:
        '200':
          description: Student found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '404':
          description: Student not found

    put:
      summary: Update a student
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the student to update
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Student'
      responses:
        '200':
          description: Student updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '404':
          description: Student not found

    delete:
      summary: Delete a student
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the student to delete
          schema:
            type: string
      responses:
        '204':
          description: Student deleted
        '404':
          description: Student not found
...
components:
  schemas:
    Student:
      type: object
      properties:
        idStudent:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        genre:
          type: string
          enum: [F, M, O]
        birthDate:
          type: string
        observation:
          type: string
        groupId:
          type: string
...